server/requirements.txt
colorama==0.4.6
iniconfig==2.0.0
lxml==4.9.3
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
python-docx==1.0.0
typing_extensions==4.8.0